// core/user-dashboard/store/auth/action-types.ts
export enum ActionTypes {
  SIGN_IN = "SIGN_IN",
  SIGN_UP = "SIGN_UP",
  SIGN_OUT = "SIGN_OUT",
  GET_ME = "GET_ME",
  VERIFY_ACCESS = "VERIFY_ACCESS",
  ENABLE_2FA = "ENABLE_2FA",
  DISABLE_2FA = "DISABLE_2FA",
  ENABLE_2FA_CONFIRMATION = "ENABLE_2FA_CONFIRMATION",
  DISABLE_2FA_CONFIRMATION = "DISABLE_2FA_CONFIRMATION",
  VERIFY_2FA = "VERIFY_2FA",
}